resolvendo bugs de emprestimo dos livros

+++ src/controllers/livro/LivroController.js
@@ -1,4 +1,5 @@
 const database = require("../../../helpers/db/database");
+const dotenv =  require('dotenv').config({ path: './config.env' });
 const nodemailer = require('nodemailer');
 const { select } = require("../../../helpers/db/database");
 // //Configuração para envio de Notificação
@@ -7,8 +8,8 @@ let transporter = nodemailer.createTransport({
     port: 465,
     secure: true, // true for 465, false for other ports
     auth:{
-        user:"colegiotaua2201@gmail.com",
-        pass:"Colegiot@u@",
+        user:process.env.USER,
+        pass:process.env.PASS,
     },
 });
             
@@ -96,22 +97,31 @@ module.exports = {
     },
     async removerLivro(req,res){
         const { id_livro } = req.params;
-        console.log(id_livro);
         const id_dono = req.headers.authorization;
 
-        let resultado = null;
+
+        let resultado = undefined;
         try {
+            let livro = await database.select('Select * from livro where id = ?', id_livro);
+
+            if (livro[0].id_usuario != id_dono) {
+                return res.status(400).json({ erro: "Você não tem autprização para efeturar esta ação" });
+            }
+
             resultado = await database.delete('Delete from livro where id = ?', id_livro);
+
+
+            if (resultado>0) {
+                return res.status(204).send();
+            } else {
+                return res.status(400).json({ erro: "Erro ao Remover Livro" });
+            }
+          
         } catch (error) {
             console.log(error);            
         }
 
-        if (resultado>0) {
-            return res.status(204).send();
-        } else {
-            return res.status(400).json({ erro: "Erro ao Remover Livro" });
-        }
-      
+   
 
     },
     async listarTodosLivros(req,res){   
@@ -145,7 +155,9 @@ module.exports = {
         }
     },
     async pedirEmprestado(req,res){
-        const {id_livro, titulo, id_requerente,id_dono} = req.body;
+        const {id_livro, titulo,id_dono} = req.body;
+        const id_requerente = req.headers.authorization;
+
         let pedido ={
             data : new Date().toJSON().slice(0, 19).replace('T', ' '),
             id_requerente : id_requerente,
@@ -166,13 +178,14 @@ module.exports = {
             }
 
             let emprestimo = await database.insert('Insert into pedido set ?',pedido );
-            console.log(emprestimo);
+            let historico = await database.insert('Insert into historico set ?',pedido );
+           
 
             let envio = await transporter.sendMail({
                 from:`${usuarios[1].nome} <${usuarios[1].email}>`,
                 to: `${usuarios[0].email}`,
-                subject: `Pedido de Emprestimo do livro ${titulo} para o usuario ${usuarios[1].nome}`,
-                text:`Olá gostaria de saber poderia pedir seu livro ${titulo} emprestado?
+                subject: `Pedido de Emprestimo do livro: ${titulo}, para o usuario: ${usuarios[1].nome}`,
+                text:`Olá gostaria de saber poderia pedir seu livro: ${titulo} emprestado?
                 Contato : ${usuarios[1].numero}`,
             })
             console.log(envio);
@@ -193,6 +206,18 @@ module.exports = {
 
         try {
             let emprestimo = await database.select('Select * from pedido where id_livro = ?', id_livro);
+            
+            const historico = emprestimo.map((pedido) =>{
+                let requisicao ={
+                 data: new Date().toJSON().slice(0, 19).replace('T', ' '),
+                 id_requerente : pedido.id_requerente,
+                 id_dono : pedido.id_dono,
+                 id_livro: id_livro,
+                 status: "Aprovado"
+                }
+                return requisicao
+            });          
+
 
             if(emprestimo[0].id_dono != id_dono){
                 return res.status(401).json({menssagem: "Voce não tem permissão para efetuar esta ação"});
@@ -202,6 +227,8 @@ module.exports = {
 
             
             await database.update('Update pedido set status = ? where id_livro = ? ',["Aprovado", id_livro] );
+            await database.insert('Insert into historico set ?',historico );
+
 
             let envio = await transporter.sendMail({
                 from:`${usuarios[0].nome} <${usuarios[0].email}>`,
@@ -212,6 +239,70 @@ module.exports = {
             })
 
             return res.status(200).json({menssagem: "Livro autorizado para emprestimo, entre em contato com o usuario"})
+        } catch (error) {
+            console.log(error);
+        } 
+        
+    },
+    async devolverLivro(req,res){
+        const { id_livro } = req.params;
+        const {titulo} = req.body;
+        const id_requerente = req.headers.authorization;
+
+        try {
+            let emprestimo = await database.select('Select * from pedido where id_livro = ?', id_livro);
+
+
+            if(emprestimo[0].id_requerente != id_requerente){
+                return res.status(401).json({menssagem: "Voce não tem permissão para efetuar esta ação"});
+            }
+
+            let  usuarios = await database.select('Select * from usuario  inner join telefone on usuario.id = telefone.id_usuario where usuario.id= ? or usuario.id = ?',[emprestimo[0].id_dono,emprestimo[0].id_requerente] );
+
+            let data = new Date().toJSON().slice(0, 19).replace('T', ' ')
+            await database.update('Update pedido set status = ?, data = ? where id_livro = ? ',["Devolvido - em analise",data, id_livro] );
+
+            let envio = await transporter.sendMail({
+                from:`${usuarios[0].nome} <${usuarios[0].email}>`,
+                to: `${usuarios[0].email}`,
+                subject: `Devolução de livro`,
+                text:`O livro ${titulo}, foi devolvido pelo usuário? ${usuarios[1].nome}, confirme na aplicação esta devolução`,
+            })
+
+            return res.status(200).json({menssagem: "Sua colicitação de devolução foi realizada com sucesso"})
+        } catch (error) {
+            console.log(error);
+        } 
+        
+    },
+    async confirmarDevolucao(req,res){
+        const { id_livro } = req.params;
+        const id_dono = req.headers.authorization;
+
+        try {
+            let emprestimo = await database.select('Select * from pedido where id_livro = ?', id_livro);
+            console.log(emprestimo[0]);
+            
+            if(emprestimo[0].id_dono != id_dono){
+                return res.status(401).json({menssagem: "Voce não tem permissão para efetuar esta ação"});
+            }
+            const historico = emprestimo.map((pedido) =>{
+                let requisicao ={
+                 data: new Date().toJSON().slice(0, 19).replace('T', ' '),
+                 id_requerente : pedido.id_requerente,
+                 id_dono : pedido.id_dono,
+                 id_livro: id_livro,
+                 status: "Devolvido"
+                }
+                return requisicao
+            });  
+
+
+            await database.delete('delete from pedido where id = ?',emprestimo[0].id );
+            await database.insert('Insert into historico set ?',historico );
+
+
+            return res.status(200).json({menssagem: "Confirmação realizada com sucesso"})
         } catch (error) {
             console.log(error);
